demo/pyproject.toml
[project]
name = "talk2ep-demo"
version = "0.1.0"
description = "Talk2EP: Natural language interface for EnergyPlus using SLIDERS RAG"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.3.33",
    "langchain>=0.3.3",
    "langchain-google-vertexai>=2.0.0",
    "langchain-openai>=0.2.2",
    "langchain-anthropic>=0.3.0",
    "mcp[cli]",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
